Add files via upload
Added the given drive.cpp and changed floats to be ints instead

+++ Task/circle.h
@@ -7,16 +7,16 @@
 
 class Circle : public Shape, public Movable {
 private:
-	float area;
-	float perimeter;
+	int area;
+	int perimeter;
 	Point origin;
 	bool is_circular;
 	std::vector<Point*> points;
-	float radius;
+	int radius;
 public:
 	Circle();
-	Circle(float radius);
-	Circle(float x, float y, float radius);
+	Circle(int radius);
+	Circle(int x, int y, int radius);
 	void initialisePoints();
 	const std::vector<Point*>& getPoints() const;
 	void addPoint(Point& point) override;
@@ -24,8 +24,8 @@ public:
 	void calculatePerimeter() override;
 	std::string toString() override;
 	void calculatePoints() override;
-	void move(float x_translation, float y_translation) override;
-	void scale(float multiplier) override;
-	void scale(float x_multiplier, float y_multiplier) override;
+	void move(int x_translation, int y_translation) override;
+	void scale(int multiplier) override;
+	void scale(int x_multiplier, int y_multiplier) override;
 	friend std::ostream& operator<<(std::ostream& out, Circle& circle);
 };
