+++ Task/movable.h
@@ -5,7 +5,7 @@
 
 class Movable {
 public:
-	virtual void move(float x_translation , float y_translation) = 0;
-	virtual void scale(float multiplier) = 0;
-	virtual void scale(float x_multiplier, float y_multiplier) = 0;
+	virtual void move(int x_translation , int y_translation) = 0;
+	virtual void scale(int multiplier) = 0;
+	virtual void scale(int x_multiplier, int y_multiplier) = 0;
 };
