+++ Task/point.cpp
@@ -4,13 +4,13 @@
 
 Point::Point() : x(0), y(0) {}
 
-Point::Point(float x, float y) : x(x), y(y) {}
+Point::Point(int x, int y) : x(x), y(y) {}
 
-std::pair<float, float> Point::getPoint() const {
+std::pair<int, int> Point::getPoint() const {
 	return std::make_pair(x, y);
 }
 
-void Point::setPoint(float new_x, float new_y) {
+void Point::setPoint(int new_x, int new_y) {
 	x = new_x;
 	y = new_y;
 }
