+++ Task/point.h
@@ -3,13 +3,13 @@
 
 class Point {
 private:
-    float x;
-    float y;
+    int x;
+    int y;
 public:
     Point();
-    Point(float x, float y);
-    std::pair<float, float> getPoint() const;
-    void setPoint(float x, float y);
+    Point(int x, int y);
+    std::pair<int, int> getPoint() const;
+    void setPoint(int x, int y);
     std::string toString();
     Point operator+(Point& other_point) const {
         return Point(x + other_point.getPoint().first, y + other_point.getPoint().second);
@@ -20,7 +20,7 @@ public:
     Point operator*(Point& other_point) const {
         return Point(x * other_point.getPoint().first, y * other_point.getPoint().second);
     }
-    Point operator*(float multiplier) const {
+    Point operator*(int multiplier) const {
         return Point(x * multiplier, y * multiplier);
     }
 };
