+++ Task/rectangle.h
@@ -7,17 +7,17 @@
 
 class Rectangle : public Shape, public Movable {
 private:
-	float area;
-	float perimeter;
+	int area;
+	int perimeter;
 	Point origin;
 	bool is_circular;
 	std::vector<Point*> points;
-	float width;
-	float height;
+	int width;
+	int height;
 public:
 	Rectangle();
-	Rectangle(float width, float height);
-	Rectangle(float x, float y, float width, float height);
+	Rectangle(int width, int height);
+	Rectangle(int x, int y, int width, int height);
 	void initialisePoints();
 	const std::vector<Point*>& getPoints() const;
 	void addPoint(Point& point) override;
@@ -25,8 +25,8 @@ public:
 	void calculatePerimeter() override;
 	std::string toString() override;
 	void calculatePoints() override;
-	void move(float x_translation, float y_translation) override;
-	void scale(float multiplier) override;
-	void scale(float x_multiplier, float y_multiplayer) override;
+	void move(int x_translation, int y_translation) override;
+	void scale(int multiplier) override;
+	void scale(int x_multiplier, int y_multiplayer) override;
 	friend std::ostream& operator<<(std::ostream& out, Rectangle& rectangle);
 };
