+++ Task/square.h
@@ -7,16 +7,17 @@
 
 class Square : public Shape, public Movable {
 private:
-	float area;
-	float perimeter;
+	
+	int area;
+	int perimeter;
 	Point origin;
 	bool is_circular;
 	std::vector<Point*> points;
-	float side;
+	int side;
 public:
 	Square();
-	Square(float side);
-	Square(float x, float y, float side);
+	Square(int side);
+	Square(int x, int y, int side);
 	void initialisePoints();
 	const std::vector<Point*>& getPoints() const;
 	void addPoint(Point& point) override;
@@ -24,8 +25,8 @@ public:
 	void calculatePerimeter() override;
 	std::string toString() override;
 	void calculatePoints() override;
-	void move(float x_translation, float y_translation) override;
-	void scale(float multiplier) override;
-	void scale(float x_multiplier, float y_multiplier) override;
+	void move(int x_translation, int y_translation) override;
+	void scale(int multiplier) override;
+	void scale(int x_multiplier, int y_multiplier) override;
 	friend std::ostream& operator<<(std::ostream& out, Square& square);
 };
